requirements.txt
# Core Dependencies
dspy
openai>=1.12.0

# Data Models & Validation
pydantic>=2.6.0
pydantic-settings>=2.0.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0

# NLP & ML
spacy>=3.7.0
transformers>=4.37.0
torch>=2.1.0
sentence-transformers>=2.3.0

# Vector Search & Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Date & Text Processing
python-dateutil>=2.8.2
phonenumbers>=8.13.0
langdetect>=1.0.9

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Redis & Caching
redis>=5.0.0
hiredis>=2.3.0

# Async Processing
celery>=5.3.0
kombu>=5.3.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# HTTP & Networking
httpx>=0.26.0
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging & Monitoring
loguru>=0.7.2
prometheus-client>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
hypothesis>=6.98.0

# Code Quality
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pylint>=3.0.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.5.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.7

# Development
ipython>=8.20.0
jupyter>=1.0.0
notebook>=7.0.0
azure-ai-documentintelligence
email-validator